+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+ - Added strip trailing slashes
+ - Added redirect trailing slashes
+
 ## 1.0.0 - 2025-09-20
 
 _Initial release_